requirements.txt
Django==2.1.1
django-ckeditor==5.6.1
django-haystack==2.8.1
django-js-asset==1.1.0
jieba==0.39
Pillow==5.2.0
pytz==2018.5
Whoosh==2.7.4
